requirements.txt
PyGObject>=3.28.3
psutil>=5.4.3
python-xlib>=0.33
